Add files via upload
Face recognition has consistently been one of the most
difficult and significant tasks in Computer Vision over the
years. Face recognition has sparked a lot of attention.
Initially, it was linked to security concerns, as automatic face
analysis might aid security services in detecting threats
Fraudulent passports, the identification of criminals or
missing children, the prevention of identity theft, and so on.